+++ data.py
@@ -1,86 +1,77 @@
-import sugartensor as tf
-import numpy as np
+import unicodedata
+from data_flow import *
 
 
 __author__ = 'namju.kim@kakaobrain.com'
 
 
-class ComTrans(object):
+# max
+_max_len = 250
 
-    def __init__(self, batch_size=32, name='train'):
 
-        # load train corpus
-        sources, targets = self._load_corpus(mode='train')
+# pre-calculated for speed-up.
+_set_size = 1456501
 
-        # to constant tensor
-        source = tf.convert_to_tensor(sources)
-        target = tf.convert_to_tensor(targets)
 
-        # create queue from constant tensor
-        source, target = tf.train.slice_input_producer([source, target])
+@sg_ventilator()
+def _vent_corpus(opt):
 
-        # create batch queue
-        batch_queue = tf.train.shuffle_batch([source, target], batch_size,
-                                             num_threads=32, capacity=batch_size*64,
-                                             min_after_dequeue=batch_size*32, name=name)
+    # load en-de parallel corpus
+    with open('asset/data/wmt_fr_en_train.msgpack', 'rb') as f:
+        en_codes, de_codes = msgpack.unpackb(f.read())
 
-        # split data
-        self.source, self.target = batch_queue
-
-        # calc total batch count
-        self.num_batch = len(sources) // batch_size
-
-        # print info
-        tf.sg_info('Train data loaded.(total data=%d, total batch=%d)' % (len(sources), self.num_batch))
-
-    def _load_corpus(self, mode='train'):
-
-        # load en-fr parallel corpus
-        from nltk.corpus import comtrans
-        als = comtrans.aligned_sents('alignment-en-fr.txt')
-
-        # make character-level parallel corpus
-        all_byte, sources, targets = [], [], []
-        for al in als:
-            src = [ord(ch) for ch in ' '.join(al.words)]  # source language byte stream
-            tgt = [ord(ch) for ch in ' '.join(al.mots)]  # target language byte stream
-            sources.append(src)
-            targets.append(tgt)
-            all_byte.extend(src + tgt)
-
-        # make vocabulary
-        self.index2byte = [0, 1] + list(np.unique(all_byte))  # add <EMP>, <EOS> tokens
-        self.byte2index = {}
-        for i, b in enumerate(self.index2byte):
-            self.byte2index[b] = i
-        self.voca_size = len(self.index2byte)
-        self.max_len = 150
-
-        # remove short and long sentence
-        src, tgt = [], []
-        for s, t in zip(sources, targets):
-            if 50 <= len(s) < self.max_len and 50 <= len(t) < self.max_len:
-                src.append(s)
-                tgt.append(t)
-
-        # convert to index list and add <EOS> to end of sentence
-        for i in range(len(src)):
-            src[i] = [self.byte2index[ch] for ch in src[i]] + [1]
-            tgt[i] = [self.byte2index[ch] for ch in tgt[i]] + [1]
-
+    # loop for each epoch
+    for en, de in zip(en_codes, de_codes):
         # zero-padding
-        for i in range(len(tgt)):
-            src[i] += [0] * (self.max_len - len(src[i]))
-            tgt[i] += [0] * (self.max_len - len(tgt[i]))
+        en += [0] * (_max_len - len(en))
+        de += [0] * (_max_len - len(de))
+        # return
+        yield en, de
 
-        # swap source and target : french -> english
-        return tgt, src
+
+class WmtFrEn(object):
+
+    def __init__(self, batch_size=16, set_type=None):
+
+        # load vocabulary
+        with open('asset/data/wmt_fr_en_dic.msgpack', 'rb') as f:
+            self.index2char, self.char2index = msgpack.unpackb(f.read())
+
+        # set voca size
+        self.voca_size = len(self.index2char)
+        self.max_len = _max_len
+
+        if set_type == 'train':
+
+            # run ventilator
+            _vent_corpus(set_type=set_type)
+
+            # get tf queue
+            en_q, fr_q = sg_tf_sinker([tf.sg_intx, tf.sg_intx])
+
+            # create batch queue
+            batch_queue = tf.train.shuffle_batch([en_q, fr_q], batch_size, shapes=[(_max_len,), (_max_len,)],
+                                                 capacity=batch_size*128, min_after_dequeue=batch_size*64,
+                                                 name='WMT_FR_EN_TRAIN')
+
+            # split data
+            self.en, self.fr = batch_queue
+
+            # calc total batch count
+            self.num_batch = _set_size // batch_size
+
+            # print info
+            tf.sg_info('WMT_FR_EN_TRAIN data loaded.(total data=%d, total batch=%d)' % (_set_size, self.num_batch))
 
     def to_batch(self, sentences):
 
         # convert to index list and add <EOS> to end of sentence
         for i in range(len(sentences)):
-            sentences[i] = [self.byte2index[ord(ch)] for ch in sentences[i]] + [1]
+            # clean white space
+            sent = ' '.join(sentences[i].lower().split())
+            # remove punctuation
+            sent = ''.join(x for x in sent if unicodedata.category(x) not in ['Pc', 'Pd', 'Ps', 'Pe', 'Pi', 'Pf', 'Po'])
+            sentences[i] = [self.char2index[ch] for ch in sent] + [1]
 
         # zero-padding
         for i in range(len(sentences)):
@@ -90,10 +81,17 @@ class ComTrans(object):
 
     def print_index(self, indices):
         for i, index in enumerate(indices):
-            str_ = ''
-            for ch in index:
-                if ch > 1:
-                    str_ += unichr(self.index2byte[ch])
-                elif ch == 1:  # <EOS>
-                    break
-            print '[%d]' % i + str_
+            str_ = self.to_str(index, split=False)
+            print(u'[%d]' % i + str_)
+
+    def to_str(self, index, split=True):
+        str_ = ''
+        for ch in index:
+            if ch > 1:
+                str_ += self.index2char[ch]
+            elif ch == 1:  # <EOS>
+                break
+        if split:
+            return str_.split(' ')
+        else:
+            return str_