+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.0.0.3 ( 2017-06-16 )
+
+Features :
+
+    - Replace comtrans corpus to WMT frech-english corpus
+    - Add BLEU score evaluation codes
+    
+
 ## 0.0.0.2 ( 2017-04-03 )
 
 Features :